+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 `ADUtils` adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+## [10.5.0] - 2020-06-18
+
+### Created
+- Create `UINavigationController.ad_popViewController(animated:completion:)`
+- Create methods to set and layout `UITableView` header and footer
+
 ## [10.4.0]
 
 ### Created
@@ -45,7 +53,7 @@ All notable changes to this project will be documented in this file.
 - Create `UIView.ad_pinToSuperview(directionalInsets:)`
 - Create `UIView.ad_pinToSuperview(usingDirectionalEdges:)`
 - Create `UIView.ad_pin(to:directionalEdges:insets:priority:)`
-- Create `UIView.ad_pin(to:directionalEdges:priority:)` 
+- Create `UIView.ad_pin(to:directionalEdges:priority:)`
 - Create `UIView.ad_pin(usingDirectionalEdges:)`
 - Create `UIView.ad_pin(to:directionalInsets:)`
 - Create `UIView.ad_pin(to:directionalEdges:)`